pyproject.toml
[tool.poetry]
name = "captain-hook"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
joke-api = "^1.0.0"
replit = "^3.1.0"
wikipedia = "^1.4.0"
youtube_dl = "^2021.6.6"
aiohttp = "^3.7.4"
discord = "^1.7.3"
python-dotenv = "^0.18.0"
requests = "^2.25.1"
DiscordUtils = "^1.3.4"
praw = "^7.3.0"
Flask = "^2.0.1"
giphy_client = "^1.0.0"
ffmpeg = "^1.4"
PyNaCl = "^1.4.0"
dbd-music = "^0.0.6"
keep-alive = "^0.0.1"
Pillow = "^8.2.0"
mal = "^0.4.0"
mal-api = "^0.4.0"
google = "^3.0.0"
beautifulsoup4 = "^4.9.3"
utils = "^1.0.1"
gevent = "^21.1.2"
asyncpg = "^0.23.0"
google-api-python-client = "^2.13.0"
psutil = "^5.8.0"
pytz = "^2021.1"
qrcode = "^7.3.1"
steam = "^1.2.0"
aiosqlite = "^0.17.0"
disputils = "^0.2.0"
humanfriendly = "^10.0"
mention = "^0.1"
pytube = "^11.0.2"
"discord.py" = "^1.7.3"
"vacefron.py" = "^1.6.3"
animec = "^0.4.2"
spotipy = "^2.22.1"
flatland-rl = "^4.0.3"
python-dateutil = "^2.9.0"
google-search = "^1.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
